bau: extract saml-test-utils
We've been shipping code we use for testing our production code within
our core libraries (Junit, Mockito, AssertJ, our fake PKI codebase etc).
We shouldn't be including code that we haven't tested or intend to use
in production when we ship code so I've extracted out this test code
into a separate help library that can be used as a test dependency  when
 we test our applications.

In some cases this is causing issues on the classpath and when we
develop our code as without some very explicit configuration in gradle
we can't be sure what version of the dependencies are included when we
compile or run tests.