BAU Downgrade dropwizard to 1.3.8
Build 179 bumped the version to 1.3.9. This caused an issue in the hub
with a bunch of integration tests which are hard to debug.

For the sake of progress with the proxy-node (which needs the hub to use
a build of saml-libs greater than 179) we're downgrading until we can
work out what's up with 1.3.9.