EID-1504 Allow eIDAS assertions to be unsigned
The eIDAS SAML profile does not require that attributes are signed. This
is different to the Verify profile which does require they are signed.

We will soon be onboarding a European country, Spain, that does not sign
its assertions. This will break saml-engine as well as the MSA.

This updates the `SamlMessageSignatureValidator` to make it possible to
skip signature validation if the message being validated is an eIDAS
attribute (determined by a flag when the validator is instantiated in
the apps).

If the attribute is signed (most countries do) we still validate that
signature and will raise if it fails - so we don't start allowing
invalid signatures. If the signature is not provided however, and the
flag is set, we just skip the validation.