ZD-3570817 duplicate matching datasets
Once every few months someone somewhere clicks the back button really quickly and generates an error in our logs; it's not really a problem and should probably be logged under WARN instead of ERROR.

Provide additional constructor for GenericHubProfileValidationSpecification that allows a Level to specified.

Use that constructor in uk.gov.ida.saml.core.validation.errors.SamlTransformationErrorFactory.duplicateMatchingDataset so that the Level is indeed WARN.

Write a test to make sure it happens.  Write a test for an arbitrary method to make sure it only happens for duplicatMatchingDataset.

Because the package names are out of sync with the folder structure, a lot of imports have changed thanks to the IDE.  Everything appears to still work despite this.