EID-1583 Revert "EID-1504 Allow eIDAS assertions to be unsigned"
This reverts all the commits included in [PR #64](https://github.com/alphagov/verify-saml-libs/pull/64).
These files have not been altered in anyway since that PR so this
reversion doesn't affect other work.

This is being done due to a security issue with the approach and a new
implementation is being worked on.