EID-1608 Check if assertions are unsigned
EID-1608 requires us to make the original SAML from the country available if the assertions are unsigned.  To determine whether an assertion is unsigned we can use the code that already exists in verify-saml-libs if we expose that functionality via a new public method.

So that's what we've done - say hello to `isAssertionUnsigned(Assertion)` - it does what it says on the tin (if the tin is made of its signature and by use of getter naming convention).