EID-1609 Re-encrypt symmetric key for receiving RP
In order to maintain trust when receiving unsigned assertions we need to provide the original signed SAML message to the RP. The RP can only decrypt encrypted assertions if it has the symmetric key to do so but in the original SAML, it's encrypted for the hub so we need to decrypt the symmetric key(s) and re-encrypt it for the receiving RP and send it in a SAML form.  This PR takes care of the decryption / re-encryption of the symmetric key(s).

Add method to assertion decrypter to enable retrieval of encrypted symmetric keys.

Add class to encrypt the symmetric keys (known as secret keys) so the RP can un-encrypt them.

Add some exception handling so that the methods are consistent with the the rest of the library.  How that can be improved or altered is open to suggestions.

Tests, of course.