EID-1609 increase test coverage
Follow the none, one and many approach to testing - with those numbers of encrypted assertions being tested. Also test Exceptions being thrown and not being thrown.

Create a few extra methods in the test framework to support the extra tests.

Add a private class that enables us to test the multiple keys case - the class implements EncryptedKey, forwards messages to an instance of a valid key for all methods except one that returns the EncryptionMethod method which returns a cunningly crafted mock that will cause a DecryptionException to be thrown.

I'm pretty confident that it's a better approach than a mock because there are lots and lots of things that would need to be When-Then-ed.