EID-1666 Add new attribute extensions for unsigned assertions
We need a way to return an entire SAML response inside another SAML
response (due to some eIDAS countries not signing their assertions). We
also need to pass the encrypted encryption keys for the assertions
inside it.

These new attributes are intended to be used to contain this information
as attributes inside a wrapper SAML response to be returned to a
service. This will then be processed by the VSP.