EID-1763 Ensure we're extracting encrypted keys
When an XML element is encrypted the information about the keying
material to decrypt it can exist in two places within the `EncryptedData`
element. It will be within its own `EncryptedKey` element.

The first is nested inside a `KeyInfo` element, which is a child of the
`EncryptedData` element.

The second is within a detached `EncryptedKey` element - that is, not
inside `KeyInfo`.

If you're interested you can (see the relevant section of the XML
encryption schema here.)[https://www.w3.org/TR/xmlenc-core1/#sec-Extensions-to-KeyInfo]

We were only set up to extract an encrypted key for the second scenario.
This commit allows us to extract a key in the first scenario.

Up until recently we'd only been extracting keys from stub-country and
from our tests. Stub-country encrypts assertions using the second method
so we hadn't caught this before. We've just run this code with Spain in
integration, who happen to use the first method.

Interestingly, (we were already validating that the encrypted assertions
used strictly one of these methods)[https://github.com/alphagov/verify-saml-security/blob/912b5b82ac2214806aebc5c92521c5b8db59c49a/src/main/java/uk/gov/ida/saml/security/validators/encryptedelementtype/EncryptionAlgorithmValidator.java#L32]
We call this validator before we get the re-encrypted keys from the
assertions. This gives me confidence that we're not going to encounter
other countries doing it other ways.

The SAML string I've included in the new test is taken from an actual
Spanish response in Integration. I've just changed the assertion cipher
data, key cipher data and key x509 in the XML for values that work with
the rest of the test setup (the test hub credentials basically), and
then converted the result to a string. This means we have a realistic
setup for an example of the second method of encryption.