EID-1763 Throw and log if missing eIDAS attributes
When we have unsigned assertions we send the SAML response and the
re-encrypted keys as attributes of a SAML assertion. In that scenario
they should both always be there.

We had an issue in the hub where the re-encrypted keys were not being
set (see previous commit). The `EidasUnsignedMatchingDatasetUnmarshaller`
was just silently returning null in this scenario instead of returning a
MatchingDataset.

The null should really be handled by the calling code
(in the MSA), but this adds some logging to help make it more obvious
what's going wrong.