EID-1762 Consolidate eIDAS hash logging
We log hashes of eIDAS attributes in the hub and the proxy-node. These
are then compared in Splunk to ensure nothing bad is happening.

The logic to do this was split over three places - the VSP, the
proxy-node and the hub. We want to consolidate the logic to ensure the
hashing is done in the same way.

The VSP extract the values from the attributes using the `MatchingDatasetToNonMatchingAttributesMapper`
which are then serialized and sent to the proxy-node using various
DTOs. This logic and DTOs have been imported to saml-lib. This allows
the hub to use the mapper to extract attributes directly from an
assertion using exactly the same logic as the VSP. The DTOs are now
available to the proxy-node to ensure that the serialized attributes are
de-serialized into the same objects.

There are some interfaces created which are implemented in the
proxy-node to ensure that the `EidasAttributesLogger` can accept certain
objects, without having to move proxy-node logic to saml-lib when it
really doesn't need to.

The `EidasAttributesLogger` class is what should be used in both the
proxy-node and the hub. It unifies the logic for extracting attributes
so that users have a consistent experience in both projects.