BAU Refactor IdaSamlBootstrap
We encountered an interesting issue where the tests were passing locally
fine, but one was failing when running in the Concourse build pipeline.

It turned out that we had to `IdaSamlBootstrap` classes, both doing very
similar things. The class registers objects and their marshallers in a
registry, allowing for for serialization.

The classes were identical, apart from one added more objects to the
registry. In that class, one of the objects `Version` was actually
overwriting an object added earlier in the class
`StringBasedMdsAttributeValue`. This is because they have the same
Qname. This has never been an issue so I've removed the
`StringBasedMdsAttributeValue` registration.

The second version of `IdaSamlBootstrap` that we had didn't include the
`Version` object. This mean that if that bootstrapping was run from that
class, it would re-overwrite the mapping for `Value` with the
marshallers for `StringBasedMdsAttributeValue`.

Now, this hasn't been an issue up to this point as the test for the
deserialization of the `Version` object was always run before the second
version of `IdaSamlBootstrap` was called, and so the re-overwriting was
never flagged.

For some reason today, the tests in Concourse have started being run in
a different order. I don't know why. It shouldn't matter what order they
are run in however. The test for the deserialisation of the `Version`
object now occurred after the re-overwrite.

This commit removes the redundant version of `IdaSamlBootstrap` so the
registering of objects only happens once and tests ca run in whatever
order without issue.