EID-1980 log Hub request id instad of pid from IDP
To mitigate against fraud, we log the hash of user attributes for an eIDAS journey in the Hub and in Proxy Node Translator. If these hashes do do match there may have been tampering of identities.

Currently, the user’s PID from the IDP is included in the hash. However, as we need to support “transient” PIDs for eIDAS, it means that we will send a different PID back to the requesting the country than what we received from the IDP. This means that we’ll be sending an identity with a different hash to what we had logged, which means that we won’t be able to tell if the identity has been tampered with in case there needs to be a fraud investigation.

To mitigate this, we will use the Proxy Node request ID generated by the VSP. This is the only common is we can use since it’s available both in the Hub and the Proxy Node.

Remove the PID string from the hash of eIDAS user attributes we log in hub saml-engine and proxy node translator. Instead log the hub request id generated by the VSP.