BAU - Utility class to encapsulate Base64 actions
The Verify / Proxy Node uses many of the different available libraries for Base64 encoding including `import org.glassfish.jersey.internal.util.Base64` which is a barrier to upgrading to Dropwizard 2.

This provides a facade to the java.util implementation so that the codebase can have one true base64 to rule them all.  Unfortunately saml-lib itself does some fairly funky stuff that includes injecting its own (overcooked?) base64 implementation so this immediately has a limitation (or rather there are more pressing things to do and I thought it would be much simpler) so that ironically, saml-lib doesn't use the base64 lib it provides at the moment.