BAU: Refresh metadata resolvers when trust-anchors change
Our current implementation of the EidasMetadataResolverRepository
periodically hits the trust-anchor endpoint and pulls the latest version
of the trust-anchors. It then either adds or removes metadata resolvers
to the internal map depending on the presence, or lack of presence, of
the trust anchors entityId in both the new and the current sets.

What it doesn't do is check if the trust-anchor for a particular
entityID has changed in the latest fetched trust-anchors. If the new
trust-anchors list has a new trust-anchor for a particular entityId,
it is not picked up on and the EidasMetadataResolverRepository continues
to use the old trust-anchor.

This change compares the newly fetched trust-anchors with the existing
ones and if they are different regenerates all of the metadata resolvers
for all the trust-anchors.

This approach is less precise than comparing both sets and only
refreshing the metadata resolvers for those whose trust-anchor has
changed (as well as removing resolvers for trust-anchors that have been
removed/adding resolvers for trust-anchors that have been added), but it
is easier to reason about and there is nothing precious about the
existing metadata resolvers. We also don't change the trust-anchors very
often.