BAU: Decrease trustAnchorMaxRefreshDelay
This changes the period that the trust-anchors are polled from 1 hour to
5 minutes. If a new trust-anchor is published it means that something is
already probably broken and we don't want to wait an hour for it. One
request every 5 minutes is not excessive.

Also, changing the default will mean that there's no need to update
config files for the MSA/VSP or the hub.