BAU: Set max refresh delay before getting eIDAS trust anchor
If there is an error with the call to `getTrustAnchors` when the MSA
starts up the logic drops straight to the `finally` block. The
`delayBeforeNextRefresh` field hasn't been explicitly set at this point
and so is the default for a long, 0.

This causes the next refresh to happen immediately - as there is still
an issue with the trust anchor this pattern repeats hundreds of times a
second. The application logs fill up and it's very hard to diagnose