HUB-912: Downgrade to Java 8
The MSA and the VSP need to stay compatible with Java 8 for the sake of
services.

saml-libn is a direct dependency of the MSA and the VSP.

saml-lib needs to be published with Java 8 byte code or the MSA and the
VSP won't be able to use it.

This pulls in version 396 of the utils libs, which have also just been
downgraded to Java 8 for the same reason.