BAU: Dropwizard -> 2, OpenSAML -> 3.4.6
This bumps dropwizard to 2, for use by the hub.

It bumps opensaml. The old version was dependany on an old version of
apach httpclient which was causing a class path clash in the hub.

It pulls in updated version of our utils libraries - these are currently
set to SNAPSHOT as there are PR's on these libs waiting to be merged.

It also adds a new class, OpenSAMLExtension. This is a junit 5 extension
which does the job of bootstrapping Ida SAML. This is used extensively
in the hub tests.