HUB-1055: Add metric on IDP SAML response
As a developer I want to know that the Hub is validating SAML responses using the "outgoing" SAML signing certificate, so that I know it is not safe to remove this certificate from metadata.

As part of the IDP SAML Certificate process, a new signing certificate is introduced to metadata for that IDP.

The "outgoing" cert is set to the "backup" position, and the Hub will validate SAML responses using the new cert, then the outgoing cert.

When there are two certs used for validation, and we successfully use the outgoing cert to validate, we can be sure that the IDP has not yet started using their new signing key. In this condition, set a Prometheus counter.